Introduce MediumDetailPresenter
